Move shared constants out to constants/lke.ts file
